client/requirements.txt
opencv-python
websocket-client
python-socketio
